DEVEX-179 more detail for the output of the isolation check
In analogy to the maven plugin, the gradle plugin will now output more information about the violations that lead to a compliance check failure. To reduce future maintenance efforts, a dependency to the maven plugin was established to reuse and reduce code. The plugin now relies on the rest API 2.0 of the fsm-dependency-checker and is configured via a ComplianceLevel instead of the IsolationLevel.