DEVEX-664 Introduce compliance-check task
The task "checkCompliance" can be used to verify that a module only
uses stable elements of the fs-isolated-runtime.jar. Internal classes
and deprecated code elements may change or get removed in the future,
so they will be reported in a parseable JUnit report.

It is possible to change the FirstSpirit version to check against
with the parameter "-PcomplianceCheckFsVersion=<version>" or setting
"firstSpirit.version" in the gradle.properties file.